requirements.txt
pandas >=0.23.1
Numpy >= 1.11.1
requests >= 2.14.2
lxml >= 4.5.1
appdirs >=1.4.4
pandas-datareader == 0.10.0
plotly == 5.10.0
tenacity == 8.1.0
cycler==0.11.0
kiwisolver==1.3.1
matplotlib==3.3.4
pillow==8.4.0
pyparsing==3.0.9